Move fixupKey in its own file: appshell_helpers
